requirements.txt
Flask==3.1.0
python-docx==1.1.2
EbookLib==0.18
reportlab==4.2.5
Pillow==10.4.0
pypdf==5.1.0
gunicorn==23.0.0
